ComQuest_Project/requirements.txt
astroid==1.4.8
backports.functools-lru-cache==1.3
bcrypt==3.1.1
boto==2.45.0
cffi==1.8.3
colorama==0.3.7
configparser==3.5.0
Django==1.7
django-braces==1.10.0
django-embed-video==1.1.0
django-storages==1.5.1
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.2
Pillow==3.4.2
pycparser==2.17
pylint==1.6.4
requests==2.12.4
six==1.10.0
virtualenv==15.0.3
wrapt==1.10.8
